% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applytoeachinlistworker.R
\name{applytoeachinlistworker}
\alias{applytoeachinlistworker}
\title{A worker function for possible parallelization in applytoeachinlist}
\usage{
applytoeachinlistworker(X = NULL, listobject = NULL,
  applyargs = NULL, applyfun = NULL, nameaftersplit = NULL,
  splitalongside = NULL, splitalongsidename = NULL,
  splitalongsidesplitter = NULL)
}
\arguments{
\item{X}{Internal - name of listobject which should be evaluated}

\item{listobject}{A named list object containing elements to which
the applyfun will be applied.}

\item{applyargs}{A named list of arguments supplied to applyfun.}

\item{applyfun}{The character name of a function to apply.}

\item{nameaftersplit}{The name of the argument of applyfun that
takes the settosplit dataframe after it has been split; this is
commonly 'x' or 'data'}

\item{splitalongside}{An optional dataframe that should be split alongside
the listobject, containing the factor variable splitalongsidename.}

\item{splitalongsidename}{The name of splitalongside after it has been split;
e.g. this is commonly something like "newdata".}

\item{splitalongsidesplitter}{The name of the factor variable in settosplit, over
which it should be split.}
}
\value{
This function returns a named list of results, having applied
  applyfun to settosplit for every level of the 'over' variable. So for
  example, result[["a"]] is the result of applyfun(data[data$over == 'a']).
}
\description{
A worker function for possible parallelization in applytoeachinlist
}
