% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_lm.R
\name{batch_lm}
\alias{batch_lm}
\title{Run a set of separate but similar stats::lm models in a single function call}
\usage{
batch_lm(data = NULL, lmargs = NULL, over = NULL)
}
\arguments{
\item{data}{The dataframe in which all the data for regressions are to be found.
This dataframe should contain a column that splits it into separate frames
over which mgcv::bam will be applied.}

\item{lmargs}{A named list of arguments to be passed to each bam. For example,
this could be bamargs=list('formula' = 'y ~ x + t', 'singular.ok' = FALSE)}

\item{over}{A character string giving the name of the column in data which is
used to split and pass subsets of the data to lm.}
}
\value{
This function returns a named list of lm objects, and possibly errors
  if lm fails to produce a model.
}
\description{
Run a set of separate but similar stats::lm models in a single function call
}
