% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAIC.batch_bam.R
\name{extractAIC.batch_bam}
\alias{extractAIC.batch_bam}
\title{Extract the AICs from batch_bam models.}
\usage{
\method{extractAIC}{batch_bam}(models = NULL)
}
\arguments{
\item{models}{This is a named list of bam models, probably resulting from
batch_bam.}
}
\value{
A data frame with columns corresponding to the outputs of extractAIC()
  called on an individual model. Because these are sometimes defined differently
  depending on the arguments originally passed to the bam models, these are
  difficult to standardize. However, typically the first numerical column will
  be the degrees of freedom and the second will be the AIC.
}
\description{
Extract the AICs from batch_bam models.
}
