% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAIC.batch_lm.R
\name{extractAIC.batch_lm}
\alias{extractAIC.batch_lm}
\title{Extract the AICs from batch_bam models.}
\usage{
\method{extractAIC}{batch_lm}(models = NULL)
}
\arguments{
\item{models}{This is a named list of lm models, probably resulting from
batch_lm.}
}
\value{
A data frame with columns corresponding to the outputs of extractAIC()
  called on an individual model.
}
\description{
Extract the AICs from batch_bam models.
}
