% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.batch_lm.R
\name{predict.batch_lm}
\alias{predict.batch_lm}
\title{Run predictions on batch_lm models}
\usage{
\method{predict}{batch_lm}(models = NULL, predictargs = NULL, over = NULL, newdata = NULL)
}
\arguments{
\item{models}{This is a named list of bam models, probably resulting from
batch_lm.}

\item{predictargs}{This is a named list of arguments for the predict function.}

\item{over}{The name of the column in newdata which contains the levels over
which the batch_lm was originally split (and should correspond to the names
of the listed models)}

\item{newdata}{The data frame to be passed to predict.lm after splitting by
over. Currently, all predictions must be run anew - i.e. we do not access
$fitted.}
}
\value{
A vector with predictions in the order of rows of newdata.
}
\description{
Run predictions on batch_lm models
}
