% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.batch_bam.R
\name{summary.batch_bam}
\alias{summary.batch_bam}
\title{Run summaries on batch_bam models}
\usage{
\method{summary}{batch_bam}(models = NULL)
}
\arguments{
\item{models}{This is a named list of bam models, probably resulting from
batch_bam.}
}
\value{
A named vector with containing model summaries
}
\description{
Run summaries on batch_bam models
}
